modified whole old code CRUD operation and has been enhansed performeance and maintainablility of the code on class library of Aplication layer

+++ Src/Computerized maintenance Logic layer/Module/AssetsManagement/clsLocations.cs
@@ -1,4 +1,5 @@
-﻿using computrized_maintenance_Data_Access.Data;
+﻿using CMMS_Api.DTO;
+using computrized_maintenance_Data_Access.Data;
 using computrized_maintenance_Data_Access.Entites.AssetsManagment;
 using computrized_maintenance_Data_Access.Misc;
 using Microsoft.EntityFrameworkCore;
@@ -8,59 +9,29 @@ namespace Computerized_maintenance_Logic_layer.Module.AssetsManagement
     /// <summary>
     /// Logic-layer class for managing Location entity operations (CRUD).
     /// </summary>
-    public sealed class clsLocations
+    public sealed class clsLocations(AppDbContext _Context)
     {
-        #region Private Fields & Singleton
-
-        private static AppDbContext? _Context;
-        private readonly static clsLocations _Instance = new();
-
-        public static clsLocations Instance
-        {
-            get
-            {
-                _Context = ClsUtility.ImplementDbContextService();
-                return _Instance;
-            }
-        }
-
-        #endregion
-
-        #region Properties
-
-        public int ID { get; set; }
-        public string LocationName { get; set; } = null!;
-
-        #endregion
-
-        #region Constructors
-
-        private clsLocations() { }
-
-        private clsLocations(int id, string locationName)
-        {
-            ID = id;
-            LocationName = locationName;
-        }
-
-        #endregion
+        
 
         #region CRUD Operations
 
         /// <summary>
-        /// Asynchronously find a Location by ID.
+        /// Retrieve a Location as data transfer object by its unique identifier.
         /// </summary>
-        public async Task<clsLocations?> FindAsync(int id)
+        /// <param name="Id">unique identifier of <see cref="Location"/></param>
+        /// <returns><see cref="LocationResponseDto"/>, otherwise <see langword="null"/></returns>
+        public async Task<LocationResponseDto?> FindAsync(int Id)
         {
             try
             {
+                if(Id < 1) return null;
+
                 var entity = await _Context.Locations
                                            .AsNoTracking()
-                                           .SingleOrDefaultAsync(x => x.ID == id);
+                                           .SingleOrDefaultAsync(x => x.ID == Id);
 
-                return entity is not null
-                    ? new clsLocations(entity.ID, entity.LocationName)
-                    : null;
+                return entity is Location ? new LocationResponseDto(entity.LocationName)
+                                          : null;
             }
             catch (Exception ex)
             {
@@ -70,18 +41,23 @@ namespace Computerized_maintenance_Logic_layer.Module.AssetsManagement
         }
 
         /// <summary>
-        /// Asynchronously add a new Location.
+        /// add new Location to the data source asynchronously
         /// </summary>
-        public async Task<bool> AddNewLocationAsync()
+        /// <param name="responseDto">Data transfer object <see cref="LocationResponseDto"/></param>
+        /// <returns><see langword="true"/> if save entity has been done, otherwise <see langword="false"/></returns>
+        public async Task<bool> AddNewLocationAsync(LocationResponseDto responseDto)
         {
             try
             {
+                if(!_IsInputDataValidate(responseDto)) return false;
+
                 var location = new Location
                 {
-                    LocationName = this.LocationName
+                    LocationName = responseDto.LocationName
                 };
 
                 await _Context.Locations.AddAsync(location);
+
                 return await _Context.SaveChangesAsync() > 0;
             }
             catch (Exception ex)
@@ -92,16 +68,21 @@ namespace Computerized_maintenance_Logic_layer.Module.AssetsManagement
         }
 
         /// <summary>
-        /// Asynchronously update an existing Location.
+        /// Update a Location into the data source asynchronously
         /// </summary>
-        public async Task<bool> UpdateLocationAsync()
+        /// <param name="requestDto">Data transfer object <see cref="LocationRequestDto"/></param>
+        /// <returns><see langword="true"/> if update entity has been done, otherwise <see langword="false"/></returns>
+        public async Task<bool> UpdateLocationAsync(LocationRequestDto requestDto)
         {
             try
             {
+                if(!_IsInputDataValidate(new LocationResponseDto(requestDto.LocationName), true, requestDto.ID)) 
+                    return false;
+
                 var result = await _Context.Locations
-                                           .Where(x => x.ID == this.ID)
+                                           .Where(x => x.ID == requestDto.ID)
                                            .ExecuteUpdateAsync(u => u
-                                           .SetProperty(p => p.LocationName, this.LocationName));
+                                           .SetProperty(p => p.LocationName, requestDto.LocationName));
 
                 return result > 0;
             }
@@ -113,29 +94,17 @@ namespace Computerized_maintenance_Logic_layer.Module.AssetsManagement
         }
 
         /// <summary>
-        /// Asynchronously delete a Location by ID.
+        /// Delete a Location from the data source asynchronously
         /// </summary>
-        public async Task<bool> DeleteLocationAsync()
+        /// <param name="ID">Unique identifier of <see cref="Location"/></param>
+        /// <returns><see langword="true"/> if Delete entity has been done, otherwise <see langword="false"/></returns>
+
+        public async Task<bool> DeleteLocationAsync(int ID)
         {
             try
             {
-                var result = await _Context.Locations
-                                           .Where(x => x.ID == this.ID)
-                                           .ExecuteDeleteAsync();
+                if (ID < 1) return false;
 
-                return result > 0;
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error deleting Location: {ex.Message}");
-                return false;
-            }
-        }
-
-        public static async Task<bool> DeleteLocationAsync(int ID)
-        {
-            try
-            {
                 var result = await _Context.Locations
                                            .Where(x => x.ID == ID)
                                            .ExecuteDeleteAsync();
@@ -150,49 +119,36 @@ namespace Computerized_maintenance_Logic_layer.Module.AssetsManagement
         }
 
         /// <summary>
-        /// Retrieve all Locations (read-only query).
+        /// retrieve all Locations as data transfer object
         /// </summary>
-        public async Task<IEnumerable<Location>> GetAllLocations()
+        /// <returns>list of <see cref="IEnumerable{LocationResponseDto}"/>, otherwise <see langword="null"/></returns>
+        public async Task<IEnumerable<LocationResponseDto>?> GetAllLocations()
         {
-            return  await _Context.Locations
-                           .Include(l => l.Asset)
-                           .AsNoTracking()
-                           .ToListAsync();
+        var list = await _Context.Locations
+                              .AsNoTracking()
+                              .Select(x => new LocationResponseDto(x.LocationName))
+                              .ToListAsync();
+
+            if (list.Count < 0) return null;
+
+            return list.AsEnumerable();
         }
 
         #endregion
 
         #region Additional Methods
 
-        /// <summary>
-        /// Asynchronously get a Location with its related Asset.
-        /// </summary>
-        //public async Task<Location?> GetLocationWithAssetAsync(int id)
-        //{
-        //    try
-        //    {
-        //        return await _Context.Locations
-        //                             .Include(l => l.Asset)
-        //                             .AsNoTracking()
-        //                             .SingleOrDefaultAsync(l => l.ID == id);
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        Console.WriteLine($"Error loading related Asset: {ex.Message}");
-        //        return null;
-        //    }
-        //}
-
+       
         #endregion
 
-        #region Destructor
-
-        ~clsLocations()
+       private bool _IsInputDataValidate(LocationResponseDto responseDto,bool IsRequest= false, int Id = 0)
         {
-            if (_Context is not null)
-                _Context.DisposeAsync();
+            if (IsRequest)
+            {
+                if (Id < 1) return false;   
+            }
+            if (string.IsNullOrEmpty(responseDto.LocationName)) return false;
+            return true;
         }
-
-        #endregion
     }
 }
